# docs/swagger.yaml
definitions:
  models.Appointment:
    properties:
      completed:
        default: false
        type: boolean
      created_at:
        type: string
      date:
        type: string
      duration:
        type: integer
      id:
        type: integer
      notes:
        type: string
      pet_id:
        type: integer
      reason:
        type: string
      updated_at:
        type: string
    required:
    - date
    - duration
    - pet_id
    - reason
    type: object
...
info:
  contact: {}
paths:
  /api/v1/appointments:
    get:
      consumes:
      - application/json
      description: Obtiene todas las citas o filtra por mascota si se especifica el
        parámetro `pet_id`.
      parameters:
      - description: ID de la mascota
        in: query
        name: pet_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Appointment'
            type: array
        "400":
          description: Formato de ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Mascota no encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Obtiene citas
      tags:
      - Appointments
    post:
      consumes:
      - application/json
      description: Crea una nueva cita
      parameters:
      - description: Datos de la cita
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Appointment'
        "400":
          description: Error en los datos enviados o fecha inválida
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Mascota no encontrada
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflicto con otra cita
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Crea una cita
      tags:
      - Appointments
  /api/v1/appointments/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina una cita existente
      parameters:
      - description: ID de la cita
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cita eliminada exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Formato de ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cita no encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Elimina cita
      tags:
      - Appointments
    get:
      consumes:
      - application/json
      description: Obtiene una cita
      parameters:
      - description: ID de la cita
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Appointment'
        "400":
          description: Formato de ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cita no encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Obtiene cita
      tags:
      - Appointments
    put:
      consumes:
      - application/json
      description: Actualiza una cita existente
      parameters:
      - description: ID de la cita
        in: path
        name: id
        required: true
        type: integer
      - description: Datos de la cita
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Appointment'
        "400":
          description: Error en los datos enviados o fecha inválida
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cita o mascota no encontrada
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflicto con otra cita
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Actualiza cita
      tags:
      - Appointments
...
  /api/v1/pets/{id}/appointments:
    get:
      consumes:
      - application/json
      description: Obtiene todas las citas de una mascota específica
      parameters:
      - description: ID de la mascota
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Appointment'
            type: array
        "400":
          description: Formato de ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Mascota no encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Obtiene citas por mascota
      tags:
      - Appointments
swagger: "2.0"